Move other structs into parser
