+++ src/packs/parser/ruby/packwerk/extractor.rs
@@ -1,5 +1,7 @@
 use crate::packs::{
-    inflector_shim::to_class_case, parser::UnresolvedReference, ProcessedFile,
+    inflector_shim::to_class_case,
+    parser::{Range, UnresolvedReference},
+    ProcessedFile,
 };
 use lib_ruby_parser::{
     nodes, traverse::visitor::Visitor, Loc, Node, Parser, ParserOptions,
@@ -19,6 +21,14 @@ struct SuperclassReference {
     pub name: String,
     pub namespace_path: Vec<String>,
 }
+
+#[derive(Debug, PartialEq, Serialize, Deserialize)]
+pub struct Definition {
+    pub fully_qualified_name: String,
+    pub location: Range,
+    pub namespace_path: Vec<String>,
+}
+
 impl UnresolvedReference {
     fn possible_fully_qualified_constants(&self) -> Vec<String> {
         if self.name.starts_with("::") {
@@ -36,33 +46,6 @@ impl UnresolvedReference {
 
         possible_constants
     }
-}
-
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
-pub struct Definition {
-    pub fully_qualified_name: String,
-    pub location: Range,
-    pub namespace_path: Vec<String>,
-}
-
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Clone, Default)]
-pub struct Range {
-    pub start_row: usize,
-    pub start_col: usize,
-    pub end_row: usize,
-    pub end_col: usize,
-}
-
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
-pub struct Location {
-    pub begin: usize,
-    pub end: usize,
-}
-
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
-pub struct LocationRange {
-    pub start: Location,
-    pub end: Location,
 }
 
 struct ReferenceCollector<'a> {
